Fix MSVC warning C4244 converting from double to DWORD
> sleep_win32.c: warning C4244: function: conversion from double to DWORD, possible loss of data
> select_win32.c: warning C4244: =: conversion from double to DWORD, possible loss of data

+++ otherlibs/unix/select_win32.c
@@ -1066,7 +1066,7 @@ CAMLprim value caml_unix_select(value readfds, value writefds, value exceptfds,
 
       if (tm >= 0.0)
         {
-          milliseconds = 1000 * tm;
+          milliseconds = (DWORD)(1000 * tm);
           DEBUG_PRINT("Will wait %d ms", milliseconds);
         }
       else
